+++ packages/wrangler/src/d1/export.ts
@@ -4,7 +4,6 @@ import { spinner, spinnerWhile } from "@cloudflare/cli/interactive";
 import chalk from "chalk";
 import { Miniflare } from "miniflare";
 import { fetch } from "undici";
-import { printWranglerBanner } from "..";
 import { fetchResult } from "../cfetch";
 import { configFileName, readConfig } from "../config";
 import { getLocalPersistencePath } from "../dev/get-local-persistence-path";
@@ -13,6 +12,7 @@ import { logger } from "../logger";
 import { APIError } from "../parse";
 import { readableRelative } from "../paths";
 import { requireAuth } from "../user";
+import { printWranglerBanner } from "../wrangler-banner";
 import { Name } from "./options";
 import { getDatabaseByNameOrBinding, getDatabaseInfoFromConfig } from "./utils";
 import type { Config } from "../config";
