+++ packages/wrangler/src/deployments.ts
@@ -10,7 +10,8 @@ import { mapBindings } from "./init";
 import { logger } from "./logger";
 import * as metrics from "./metrics";
 import { requireAuth } from "./user";
-import { getScriptName, printWranglerBanner } from ".";
+import { getScriptName } from "./utils/getScriptName";
+import { printWranglerBanner } from "./wrangler-banner";
 import type { Config } from "./config";
 import type { WorkerMetadataBinding } from "./deployment-bundle/create-worker-upload-form";
 import type { ServiceMetadataRes } from "./init";
