+++ packages/wrangler/src/pages/secret/index.ts
@@ -5,13 +5,13 @@ import { getConfigCache } from "../../config-cache";
 import { findWranglerConfig } from "../../config/config-helpers";
 import { confirm, prompt } from "../../dialogs";
 import { FatalError } from "../../errors";
-import { printWranglerBanner } from "../../index";
 import isInteractive from "../../is-interactive";
 import { logger } from "../../logger";
 import * as metrics from "../../metrics";
 import { parseBulkInputToObject } from "../../secret";
 import { requireAuth } from "../../user";
 import { readFromStdin, trimTrailingWhitespace } from "../../utils/std";
+import { printWranglerBanner } from "../../wrangler-banner";
 import { PAGES_CONFIG_CACHE_FILENAME } from "../constants";
 import { EXIT_CODE_INVALID_PAGES_CONFIG } from "../errors";
 import type { Config } from "../../config";
