+++ packages/wrangler/src/versions/secrets/list.ts
@@ -2,9 +2,9 @@ import { fetchResult } from "../../cfetch";
 import { configFileName } from "../../config";
 import { createCommand } from "../../core/create-command";
 import { UserError } from "../../errors";
-import { getLegacyScriptName } from "../../index";
 import { logger } from "../../logger";
 import { requireAuth } from "../../user";
+import { getLegacyScriptName } from "../../utils/getLegacyScriptName";
 import { fetchDeploymentVersions, fetchLatestDeployment } from "../api";
 import type { VersionDetails } from ".";
 import type { ApiVersion, VersionCache } from "../types";
