+++ packages/wrangler/src/versions/secrets/put.ts
@@ -3,9 +3,9 @@ import { configFileName } from "../../config";
 import { createCommand } from "../../core/create-command";
 import { prompt } from "../../dialogs";
 import { UserError } from "../../errors";
-import { getLegacyScriptName } from "../../index";
 import { logger } from "../../logger";
 import { requireAuth } from "../../user";
+import { getLegacyScriptName } from "../../utils/getLegacyScriptName";
 import { readFromStdin, trimTrailingWhitespace } from "../../utils/std";
 import { copyWorkerVersionWithNewSecrets } from "./index";
 import type { WorkerVersion } from "./index";
